pyproject.toml
[project]
name = "rag-lab"
version = "0.2.0"
description = "Graduate data modeling lab: Postgres + document store (MongoDB/TinyDB) with FastAPI + Streamlit"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "psycopg>=3.1",
  "pydantic>=2.5",
  "pymongo>=4.6",
  "tinydb>=4.8",
  "streamlit>=1.30",
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[tool.uv]
dev-dependencies = [
  "ipykernel>=6.29",
  "jupyter>=1.0",
]
